Downstream image and script changes
- Use VMware repository for base image
- Do not download OVS source code
- Do not install IPsec related packages by default
- Remove unnecessary OVS patches from apply-patches.sh
- Configure iptables-wrapper for Ubuntu image

2023-12-26:
- Remove whereabouts per upstream change

2024-07-22:
- Fix IPsec build error caused by incorrect image tag


Note: this should be squashed to 0a80591 on the next rebase


Install iproute2 in Ubuntu Image

antctl support bundle needs the ip command.

Add Dockerfile for Debian builds

Add support for building OVS image and Antrea images with
Debian as base image.


Add Dockerfile for Photon builds

Add support for building OVS image and Antrea images with
Photon as base image.


Add Dockerfile for Photon builds

Add support for building OVS image and Antrea images with
Photon as base image.


Adjust UBI Dockerfile for OVS and Antrea


Support building Photon locally


Support IPSec in Photon base image


Install python3-curses in OVS photon

Install gzip in Photon to fix the issue of logrotate

Logrotate uses gzip to compress the log files by default.
Added an e2e test case to prevent such issues in the future.

Add net-tools to install the hostname command to prevent
errors from ovs-ctl.

Fixes: https://bugzilla.eng.vmware.com/show_bug.cgi?id=3312394


Install Suricata in Debian images


Update Photon openvswitch.spec to 2.17.0

Update openvwitch.spec for Photon


Define OVS version in photon SPEC file


Build different images for antrea-agent and antrea-controller

Build debian image for flow-aggregator

Support building flow-aggregator ubi image


Adapt to Upstream Dockerfile for flow-aggregaror

Install ca-certificates in flow-aggregator. This is for supporting
the S3 uploader.

Build and install Suricata in Photon images


Support installing Suricata from local packages


Use broadcom mirror repo for docker images

+++ hack/verify-helm.sh
@@ -32,6 +32,7 @@ verify_helm() {
         #  - v3.8.1
         #  - v3.8.1+g5cb9af4
         helm_version="${helm_version%+*}"
+        helm_version="${helm_version%+*}"
         if [ "${helm_version}" == "${_HELM_VERSION}" ]; then
             # If version is exact match, stop here.
             echo "$helm"
@@ -62,11 +63,16 @@ verify_helm() {
     esac
     
     >&2 echo "Installing helm"
-    local helm_url="https://get.helm.sh/helm-${_HELM_VERSION}-${ostype}-${arch}.tar.gz"
-    curl -sLo helm.tar.gz "${helm_url}" || return 1
     mkdir -p "$_BINDIR" || return 1
-    tar -xzf helm.tar.gz -C "$_BINDIR" --strip-components=1 "${ostype}-${arch}/helm" || return 1
-    rm -f helm.tar.gz
+    if [ -z $1 ]; then
+        local helm_url="https://get.helm.sh/helm-${_HELM_VERSION}-${ostype}-${arch}.tar.gz"
+        curl -sLo helm.tar.gz "${helm_url}" || return 1
+        tar -xzf helm.tar.gz -C "$_BINDIR" --strip-components=1 "${ostype}-${arch}/helm" || return 1
+        rm -f helm.tar.gz
+    else
+        cp $1 "$_BINDIR"
+        chmod +x "$_BINDIR"/helm
+    fi
     helm="$_BINDIR/helm"
     echo "$helm"
     return 0
